third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/xmlhttprequest-responseXML-nopartial-expected.txt

PASS xhr.responseXML is non-null.
